requirements.txt
attrs==20.3.0
aws-cdk.assets==1.98.0
aws-cdk.aws-apigateway==1.98.0
aws-cdk.aws-apigatewayv2==1.98.0
aws-cdk.aws-applicationautoscaling==1.98.0
aws-cdk.aws-autoscaling==1.98.0
aws-cdk.aws-autoscaling-common==1.98.0
aws-cdk.aws-autoscaling-hooktargets==1.98.0
aws-cdk.aws-batch==1.98.0
aws-cdk.aws-certificatemanager==1.98.0
aws-cdk.aws-cloudformation==1.98.0
aws-cdk.aws-cloudfront==1.98.0
aws-cdk.aws-cloudwatch==1.98.0
aws-cdk.aws-cloudwatch-actions==1.98.0
aws-cdk.aws-codebuild==1.98.0
aws-cdk.aws-codecommit==1.98.0
aws-cdk.aws-codeguruprofiler==1.98.0
aws-cdk.aws-codepipeline==1.98.0
aws-cdk.aws-cognito==1.98.0
aws-cdk.aws-ec2==1.98.0
aws-cdk.aws-ecr==1.98.0
aws-cdk.aws-ecr-assets==1.98.0
aws-cdk.aws-ecs==1.98.0
aws-cdk.aws-efs==1.98.0
aws-cdk.aws-elasticloadbalancing==1.98.0
aws-cdk.aws-elasticloadbalancingv2==1.98.0
aws-cdk.aws-events==1.98.0
aws-cdk.aws-events-targets==1.98.0
aws-cdk.aws-globalaccelerator==1.98.0
aws-cdk.aws-iam==1.98.0
aws-cdk.aws-kinesis==1.98.0
aws-cdk.aws-kinesisfirehose==1.98.0
aws-cdk.aws-kms==1.98.0
aws-cdk.aws-lambda==1.98.0
aws-cdk.aws-logs==1.98.0
aws-cdk.aws-route53==1.98.0
aws-cdk.aws-route53-targets==1.98.0
aws-cdk.aws-s3==1.98.0
aws-cdk.aws-s3-assets==1.98.0
aws-cdk.aws-sam==1.98.0
aws-cdk.aws-secretsmanager==1.98.0
aws-cdk.aws-servicediscovery==1.98.0
aws-cdk.aws-signer==1.98.0
aws-cdk.aws-sns==1.98.0
aws-cdk.aws-sns-subscriptions==1.98.0
aws-cdk.aws-sqs==1.98.0
aws-cdk.aws-ssm==1.98.0
aws-cdk.aws-stepfunctions==1.98.0
aws-cdk.cloud-assembly-schema==1.98.0
aws-cdk.core==1.98.0
aws-cdk.custom-resources==1.98.0
aws-cdk.cx-api==1.98.0
aws-cdk.region-info==1.98.0
beautifulsoup4==4.9.3
cattrs==1.4.0
certifi==2020.12.5
chardet==4.0.0
constructs==3.3.71
idna==2.10
iniconfig==1.1.1
# Editable Git install with no remote (job-trigger-lambda==0.0.1)
-e /Users/it-ops/Documents/java_dev/workspaces/workspace-projects/aws/job-trigger-lambda-cdk/job_trigger_lambda
jsii==1.28.0
lxml==4.6.2
multitasking==0.0.9
numpy==1.19.5
packaging==20.9
pandas==1.2.0
pluggy==0.13.1
publication==0.0.3
py==1.10.0
pyparsing==2.4.7
pytest==6.2.3
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.1
selenium==3.141.0
six==1.15.0
soupsieve==2.1
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.2
urlopen==1.0.0
